requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
BingImageCreator==0.5.0
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
decorator==5.1.1
Deprecated==1.2.14
EdgeGPT==0.13.2
exceptiongroup==1.1.2
extra-streamlit-components==0.1.56
Faker==19.3.0
favicon==0.7.0
firebase-admin==6.2.0
Flask==2.3.2
fonttools==4.42.0
frozenlist==1.4.0
gcloud==0.18.3
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-api-python-client==2.96.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
h11==0.14.0
htbuilder==0.6.1
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jwcrypto==1.5.0
kiwisolver==1.4.4
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
more-itertools==10.1.0
msgpack==1.0.5
multidict==6.0.4
numpy==1.25.2
oauth2client==4.1.3
outcome==1.2.0
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pydeck==0.8.0
Pygments==2.16.1
PyJWT==2.8.0
pymdown-extensions==10.1
Pympler==1.0.1
pyparsing==3.0.9
Pyrebase4==4.7.1
PySocks==1.7.1
python-dateutil==2.8.2
python-jwt==4.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.29.0
requests-toolbelt==0.10.1
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
selenium==4.11.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.4.1
st-annotated-text==4.0.0
streamlit==1.25.0
streamlit-authenticator==0.2.2
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-elements==0.1.0
streamlit-embedcode==0.1.2
streamlit-extras==0.3.0
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-option-menu==0.3.6
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uritemplate==4.1.1
urllib3==1.26.16
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.6
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
zipp==3.16.2
